move Load... functions only used in tests to tests.go
