+++ config/configload/tests.go
@@ -29,3 +29,24 @@ func loadTestContents(tcs []testContent) (*config.Couper, error) {
 
 	return bodiesToConfig(parsedBodies, srcs, "")
 }
+
+func LoadFile(file, env string) (*config.Couper, error) {
+	return LoadFiles([]string{file}, env)
+}
+
+func LoadBytes(src []byte, filename string) (*config.Couper, error) {
+	return LoadBytesEnv(src, filename, "")
+}
+
+func LoadBytesEnv(src []byte, filename, env string) (*config.Couper, error) {
+	hclBody, err := parser.Load(src, filename)
+	if err != nil {
+		return nil, err
+	}
+
+	if err = validateBody(hclBody, false); err != nil {
+		return nil, err
+	}
+
+	return bodiesToConfig([]*hclsyntax.Body{hclBody}, [][]byte{src}, env)
+}
